refactor(cli): Lazy load locator and daemon modules if needed
